pyproject.toml
[project]
name = "studiofinder-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.5.4",
    "duckduckgo-search>=8.0.2",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.32.0",
]
